+++ legadilo/reading/models/article.py
@@ -135,6 +135,7 @@ class ArticleFullTextSearchQuery(ArticleSearchQuery):
     search_type: constants.ArticleSearchType = constants.ArticleSearchType.PLAIN
     order: constants.ArticleSearchOrderBy = constants.ArticleSearchOrderBy.RANK_DESC
     linked_with_feeds: frozenset[int] = frozenset()
+    external_tags_to_include: frozenset[str] = frozenset()
 
     @property
     def order_by(self) -> str:  # noqa: PLR0911 Too many return statements
@@ -301,6 +302,13 @@ class ArticleQuerySet(models.QuerySet["Article"]):
             .prefetch_related(_build_prefetch_article_tags())
             .default_order_by()
         )
+
+    def for_external_tags(self, tags: list[str]) -> Self:
+        filters = models.Q()
+        for tag in tags:
+            filters |= models.Q(external_tags__icontains=tag)
+
+        return self.filter(filters)
 
     def for_feed(self) -> Self:
         return (
@@ -703,6 +711,9 @@ class ArticleManager(models.Manager["Article"]):
 
         if search_query.linked_with_feeds:
             articles_qs = articles_qs.filter(feeds__id__in=search_query.linked_with_feeds)
+
+        if search_query.external_tags_to_include:
+            articles_qs = articles_qs.for_external_tags(search_query.external_tags_to_include)
 
         if search_query.search_type == constants.ArticleSearchType.URL:
             articles_qs = articles_qs.for_url_search([search_query.q])