+++ legadilo/reading/tests/test_views/test_search_views.py
@@ -20,6 +20,7 @@ from http import HTTPStatus
 from urllib.parse import urlencode
 
 import pytest
+from django.http import QueryDict
 from django.urls import reverse
 
 from legadilo.conftest import assert_redirected_to_login_page
@@ -29,40 +30,19 @@ from legadilo.reading import constants
 from legadilo.reading.models import ArticleTag
 from legadilo.reading.tests.factories import ArticleFactory, TagFactory
 from legadilo.reading.views.search_views import SearchForm
+from legadilo.utils.http_utils import dict_to_query_dict
 from legadilo.utils.testing import AnyOfType
 
 
 class TestSearchForm:
     def test_without_data(self):
-        form = SearchForm({}, tag_choices=[], feeds_qs=Feed.objects.none())
+        form = SearchForm(QueryDict(), tag_choices=[], feeds_qs=Feed.objects.none())
 
         assert form.is_valid()
 
     def test_with_only_q(self):
-        form = SearchForm({"q": "Claudius"}, tag_choices=[], feeds_qs=Feed.objects.none())
-
-        assert form.is_valid()
-        assert form.cleaned_data == {
-            "q": "Claudius",
-            "search_type": constants.ArticleSearchType.PLAIN,
-            "order": constants.ArticleSearchOrderBy.RANK_DESC,
-            "read_status": constants.ReadStatus.ALL,
-            "favorite_status": constants.FavoriteStatus.ALL,
-            "for_later_status": constants.ForLaterStatus.ALL,
-            "articles_max_age_value": None,
-            "articles_max_age_unit": constants.ArticlesMaxAgeUnit.UNSET,
-            "articles_reading_time": None,
-            "articles_reading_time_operator": constants.ArticlesReadingTimeOperator.UNSET,
-            "include_tag_operator": constants.ReadingListTagOperator.ALL,
-            "tags_to_include": [],
-            "exclude_tag_operator": constants.ReadingListTagOperator.ALL,
-            "tags_to_exclude": [],
-            "linked_with_feeds": AnyOfType(Feed.objects.none()),
-        }
-
-    def test_q_cleaning(self):
         form = SearchForm(
-            {"q": "<span>Claudius</span>"}, tag_choices=[], feeds_qs=Feed.objects.none()
+            dict_to_query_dict({"q": "Claudius"}), tag_choices=[], feeds_qs=Feed.objects.none()
         )
 
         assert form.is_valid()
@@ -81,11 +61,43 @@ class TestSearchForm:
             "tags_to_include": [],
             "exclude_tag_operator": constants.ReadingListTagOperator.ALL,
             "tags_to_exclude": [],
+            "external_tags_to_include": [],
+            "linked_with_feeds": AnyOfType(Feed.objects.none()),
+        }
+
+    def test_q_cleaning(self):
+        form = SearchForm(
+            dict_to_query_dict({"q": "<span>Claudius</span>"}),
+            tag_choices=[],
+            feeds_qs=Feed.objects.none(),
+        )
+
+        assert form.is_valid()
+        assert form.cleaned_data == {
+            "q": "Claudius",
+            "search_type": constants.ArticleSearchType.PLAIN,
+            "order": constants.ArticleSearchOrderBy.RANK_DESC,
+            "read_status": constants.ReadStatus.ALL,
+            "favorite_status": constants.FavoriteStatus.ALL,
+            "for_later_status": constants.ForLaterStatus.ALL,
+            "articles_max_age_value": None,
+            "articles_max_age_unit": constants.ArticlesMaxAgeUnit.UNSET,
+            "articles_reading_time": None,
+            "articles_reading_time_operator": constants.ArticlesReadingTimeOperator.UNSET,
+            "include_tag_operator": constants.ReadingListTagOperator.ALL,
+            "tags_to_include": [],
+            "exclude_tag_operator": constants.ReadingListTagOperator.ALL,
+            "tags_to_exclude": [],
+            "external_tags_to_include": [],
             "linked_with_feeds": AnyOfType(Feed.objects.none()),
         }
 
     def test_q_cleaning_result_too_small_after_cleaning(self):
-        form = SearchForm({"q": "<span>C</span>"}, tag_choices=[], feeds_qs=Feed.objects.none())
+        form = SearchForm(
+            dict_to_query_dict({"q": "<span>C</span>"}),
+            tag_choices=[],
+            feeds_qs=Feed.objects.none(),
+        )
 
         assert not form.is_valid()
         assert form.errors == {
@@ -94,11 +106,11 @@ class TestSearchForm:
 
     def test_advanced_values_set_unit_and_operator_unset(self):
         form = SearchForm(
-            {
+            dict_to_query_dict({
                 "q": "Claudius",
                 "articles_max_age_value": 12,
                 "articles_reading_time": 12,
-            },
+            }),
             tag_choices=[],
             feeds_qs=Feed.objects.none(),
         )
@@ -113,11 +125,11 @@ class TestSearchForm:
 
     def test_advanced_values_unset_unit_and_operator_set(self):
         form = SearchForm(
-            {
+            dict_to_query_dict({
                 "q": "Claudius",
                 "articles_max_age_unit": constants.ArticlesMaxAgeUnit.WEEKS.value,
                 "articles_reading_time_operator": constants.ArticlesReadingTimeOperator.LESS_THAN.value,  # noqa: E501
-            },
+            }),
             tag_choices=[],
             feeds_qs=Feed.objects.none(),
         )
@@ -132,7 +144,7 @@ class TestSearchForm:
 
     def test_filled(self):
         form = SearchForm(
-            {
+            dict_to_query_dict({
                 "q": "Claudius",
                 "search_type": constants.ArticleSearchType.PHRASE,
                 "order": constants.ArticleSearchOrderBy.READ_AT_ASC,
@@ -147,8 +159,9 @@ class TestSearchForm:
                 "tags_to_include": ["some-tag"],
                 "exclude_tag_operator": constants.ReadingListTagOperator.ANY.value,
                 "tags_to_exclude": ["other-tag"],
+                "external_tags_to_include": ["Some word"],
                 "linked_with_feeds": [],
-            },
+            }),
             tag_choices=[("some-tag", "Some tag"), ("other-tag", "Other tag")],
             feeds_qs=Feed.objects.none(),
         )
@@ -169,6 +182,7 @@ class TestSearchForm:
             "tags_to_include": ["some-tag"],
             "exclude_tag_operator": constants.ReadingListTagOperator.ANY,
             "tags_to_exclude": ["other-tag"],
+            "external_tags_to_include": ["Some word"],
             "linked_with_feeds": AnyOfType(Feed.objects.none()),
         }
 
@@ -241,6 +255,19 @@ class TestSearchView:
         assert response.template_name == "reading/search.html"
         assert response.context_data["search_form"].is_valid()
         assert response.context_data["articles"] == [article_with_tag_to_include]
+        assert response.context_data["total_results"] == 1
+
+    def test_search_with_external_tags(self, user, logged_in_sync_client):
+        article = ArticleFactory(user=user, external_tags=["Poésie"])
+
+        response = logged_in_sync_client.get(
+            self.url, data={"external_tags_to_include": ["Poésie"]}
+        )
+
+        assert response.status_code == HTTPStatus.OK
+        assert response.template_name == "reading/search.html"
+        assert response.context_data["search_form"].is_valid()
+        assert response.context_data["articles"] == [article]
         assert response.context_data["total_results"] == 1
 
     def test_search_with_feeds(self, user, logged_in_sync_client):