+++ legadilo/utils/tests/test_http_utils.py
@@ -1,0 +1,25 @@
+# Legadilo
+# Copyright (C) 2023-2025 by Legadilo contributors.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+# SPDX-License-Identifier: AGPL-3.0-or-later
+
+from django.http import QueryDict
+
+from legadilo.utils.http_utils import dict_to_query_dict
+
+
+def test_dict_to_query_dict():
+    assert dict_to_query_dict({"a": 1, "b": [2, 3]}) == QueryDict("a=1&b=2&b=3")